% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whr-data.R
\docType{data}
\name{whr_raw}
\alias{whr_raw}
\title{World Happiness Report 2025 demonstration data}
\format{
A data frame with 145 countries and 6 columns:
\describe{
\item{WELL}{Life evaluation, three-year average.}
\item{GDP}{Contribution of log GDP per capita.}
\item{SOC}{Contribution of social support.}
\item{LIFE}{Contribution of healthy life expectancy.}
\item{FREE}{Contribution of freedom to make life choices.}
\item{GEN}{Contribution of generosity.}
}

Country names are stored as row names. The explanatory-factor columns are
World Happiness Report decomposition components, not raw survey responses.
}
\source{
World Happiness Report 2025, data for Figure 2.1, used with
permission. See \url{https://www.worldhappiness.report/data-sharing/}.

Helliwell, J. F., Layard, R., Sachs, J. D., De Neve, J.-E., Aknin, L. B.,
& Wang, S. (Eds.). (2025). \emph{World Happiness Report 2025}. University of
Oxford: Wellbeing Research Centre.
}
\usage{
whr_raw
}
\description{
A compact country-level data set derived from the World Happiness Report
2025 data for Figure 2.1. It is included to support qcaERT demonstrations
based on a real QCA workflow.
}
\examples{
data(whr_raw)
head(whr_raw)

}
\keyword{datasets}
